rename config
