+++ src/index.ts
@@ -11,7 +11,7 @@ export {
     optionsSchema,
     projectSchema,
     configSchema,
-} from "./config/parseConfig";
+} from "./000-config/parseConfig";
 export { AstSorter } from "./grammar/sort";
 export { AstRenamer } from "./grammar/rename";
 export { AstHasher } from "./grammar/hash";
