+++ src/verify.ts
@@ -1,6 +1,6 @@
 import normalize from "path-normalize";
 import { Cell } from "@ton/core";
-import { Config, Options } from "./config/parseConfig";
+import { Config, Options } from "./000-config/parseConfig";
 import { ILogger, Logger } from "./logger";
 import { PackageFileFormat, run } from "./main";
 import { fileFormat } from "./packaging/fileFormat";
